Testing/Doc
Add a test that reproduces the SplineCurve geometry using a file instead of having the spline data in the control file. The test uses the same BenchmarkData because the results should be the same. Remove an extraneous print. Update the documentation to include the spline from file. Change the names on the 2D example to remove the 3D.