With bi cubics (#35)
* DataFileTopography upgrade

Create the biCubic interpolation class. Use it for computing the hight and the gaussian curvature.

* Add Testing

for Bicubic interpolation

* Sizing flag

Only compute bicubic for gaussian curvature if sizing is ON.

* Add Benchmarking files

Benchmark files for "SeaMountCubic" are added.

* Update Makefile

Add BiCubicClass to makefile.

* Documentation/Examples/Code update

Update the documentation to include the sizing option on the topography. Add the change to the Pond example, and change the description in the examples.md table. Get rid of a development block that limited the size of the curvature.

* add a figure caption for new three dimensional docs example

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>