Topology data edge considerations
Use the first derivatives from one point in in the BiCubic class to define edge derivatives. In the dataFile topography class, do the same thing for the Gaussian curvature. Also, compute and save the bounding box, with a little slop, and post a warning exception if the sizer tries to go outside the region where the data is defined.