Topography resolution (#33)
* add capability to read in bottom information from a file and create an interpolant

* add test control file for new feature

* update Makefile with new module

* betters comments

* hook the new feature into the SMModel

* try to use INQUIRE and throw exception but there is an issue

* Add USE ErrorTypesModule to import routine for throwing exceptions

* can now throw a fatal error if the topography data file is invalid

* add new global constant for a deafult file path length

* properly pick the file unit for readin. Adjust error exception call

* remove bilinear interpolant and reorder functions

* store the matrix in compressed column format to save on code length and (slightly) on time

* Enable boundary interpolant for extruded meshes on the first layer

* This ensures that bottom topography contains interpolant information
  for the boundary specification. However, this change may introduce
  unnecessary interpolant information for extruded models that do not
  require interpolant data.

* Fix refinement bug

Fix the problem where overrefinement happens near a straight boundary. This was due to the dz component not being ignored.

* Allow topography data to be rectangular

* The topography IO allowed only square grids to be specified to input
  topography. To permit topographic grids with domains with different
  side lengths, `nnodes` has been transitioned to `nx,ny`. Necessary
changes to the data structure and routines where nnodes was referenced
have been applied in this commit.

* The MtStHelens example file has been adjusted to list `nx, ny` just
  after the first header line.

* Enable bottom cell sidewall interpolants in output mesh file

* For cells that are adjacent to topography, the bottom face, and four
  sidewalls need to include interpolant data in the mesh output file.
  This ensures that the bottom edges of the side walls conform to the
  bottom face of the bottom cells.

* Update ABAQUS_IceCreamCone.control

Add \n to end of the file.

* spacing adjustments

* name change of new class and data type

* Update SMTopographyClass.f90

Add gaussian curvature procedure to base class. Uses finite differences for derivatives, so can be used by any subclass.

* Add topography sizing

Fully coded, now needs testing.

* Update SMTopographyClass.f90

Account for curvature being negative.

* Update Sizer.f90

Bail from topography sizing loop when the size is enough to trigger a refinement.

* 3DTestingAndTests

Added 3D testing to code and four test cases.

* Add Flag to benchmarks

Ad the Quad mesh flag - 1 to the 2D benchmark data files.

* Update MeshQualityAnalysis.f90

Add more comments about addition of 3D shape measure.

* fix incorrect size in test setup array

* update St Helens example. Omit the tec and inp files due to size

* add smaller bottom extrusion from file example and a corresponding test

* added documentation on how to add a new test case

* update the 3d examples list

* add description of gridded data mesh file format to the docs

* added example code of setting a TOPOGEPHY file for readin

* Add Finals

Add FINAL declarations for classes where they had been forgotten.

* Remove extra routine

Remove addStringToDictionary from DataFileTopographyClass used in the parent class. Make that routine private in parent.

* Create GaussianCurvature.f90

Add in gaussian curvature file.

* Update Makefile

Add gaussianCurvature compilation.

* Update MeshingTests.f90

Add test case for topography controlled sizing.

* Add Benchmark files

Add benchmark files for new SeaMount test with topography based refinement.

* Update Makefile

Fix makefile dependency

* add SeaMount to the external test list

* Gaussian Curvature Tests

Add tests for the Gaussian curvature module.

* Gaussian Curvature update

Use Gaussian curvature routine to compute Gaussian curvature. Function version not used to save some function evaluations.

* Add reference to sizing parameter in TOPOGRAPHY blocks documentation

* Revert "Add reference to sizing parameter in TOPOGRAPHY blocks documentation"

This reverts commit 6e9af0d1677500fcb414c3b9788e76302b0799b7.

* With bi cubics (#35)

* DataFileTopography upgrade

Create the biCubic interpolation class. Use it for computing the hight and the gaussian curvature.

* Add Testing

for Bicubic interpolation

* Sizing flag

Only compute bicubic for gaussian curvature if sizing is ON.

* Add Benchmarking files

Benchmark files for "SeaMountCubic" are added.

* Update Makefile

Add BiCubicClass to makefile.

* Documentation/Examples/Code update

Update the documentation to include the sizing option on the topography. Add the change to the Pond example, and change the description in the examples.md table. Get rid of a development block that limited the size of the curvature.

* add a figure caption for new three dimensional docs example

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>

* Topology data edge considerations

Use the first derivatives from one point in in the BiCubic class to define edge derivatives. In the dataFile topography class, do the same thing for the Gaussian curvature. Also, compute and save the bounding box, with a little slop, and post a warning exception if the sizer tries to go outside the region where the data is defined.

* Create ApproximationOfGaussianCurvature.md

Add discussion of how the Gaussian curvature is computed.

* ApproximationOfGaussianCurvature.md

Equation editing. Renders properly in VSCode with gitlab settings.

* Fix typos in docs

Just a couple of typos noticed and fixed.

* Documentation update

Fix some typos in the 3D doc. Add a link in the assumptions doc to the Gaussian curvature Doc. The actual link URL needs to be checked out.

* Update MeshProject.f90

Change the missing MODEL block warning to say that a Cartesian mesh will be generated.

* fix broken link and adjust math envirnoments. Docs and LaTeX build correctly

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>
Co-authored-by: Joe Schoonover <joe@fluidnumerics.com>