Bicubic and Evaluator mods
Change the leftIndex function to return 1 if the test point is too small and N-1 if it is too large. Explicitly use only the first character for a one character string in the equation evaluator to eliminate the string truncation warning. Return the bottom topography test to the test suite