PR for uninitializeds (#42)
* Update EquationEvaluatorClass.f90

initialize firstCharacterIsDigit

* Code Cleanup

Remove unused variables. Ensure input files are closed before moving to another benchmark. Removed bottomFromFile test to see if that is where the problem is coming from.

* Bicubic and Evaluator mods

Change the leftIndex function to return 1 if the test point is too small and N-1 if it is too large. Explicitly use only the first character for a one character string in the equation evaluator to eliminate the string truncation warning. Return the bottom topography test to the test suite