Implement subdivision check (#38)
* Implement subdivision check

Finds the number of refinements necessary to resolve boundary curves. If that number is larger than maxLevelLimit (currently 8, for a size ratio of 256) then a fatal error is posted with a message to rerun with the sLimit flag. Plus a warning to think before doing it.

* Add check for level

Check also if the level is hit during the refinement process. this can happen with refinement regions or bottom topography.

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>