Remove chain requirement for inner boundaries
Inner boundary curves no longer have to be enclosed in a CHAIN block. A chain is only needed if multiple curves are connected to define an inner boundary. If a single curve defines an inner boundary, simply use that in place of a chain. If multiple curves make up the boundary, include them in a CHAIN.