Update SMMeshClass.f90
Modify logic on node release to remove nodes from the node list if the node list is the only object that references it. Also manually remove all the linked lists from the boundaryEdgesArray before releasing it. This seems to avoid an issue where waiting until the finalizer does it causes a crash in the interface routines. It also fixes an issue where unreferenced nodes appeared in the node list in the mesh file.