Implement interface boundaries (#53)
* CurveType array allocation fix

Move curveType into the mesh class to save after the mesh has been generated.

* Cleanup interface curve code

Some basic cleanup including zeroing the node acceleration and removing a redundant subroutine. Code still crashes on final smooth with interface boundaries.

* Just typos

at this point

* Update SMMeshObjects.f90

Remove unused procedure.

* Move element deletion from interface code

All cleanup is moved out of the interface code and moved into the MeshCleaner code.

* Update SMMeshClass.f90

Modify logic on node release to remove nodes from the node list if the node list is the only object that references it. Also manually remove all the linked lists from the boundaryEdgesArray before releasing it. This seems to avoid an issue where waiting until the finalizer does it causes a crash in the interface routines. It also fixes an issue where unreferenced nodes appeared in the node list in the mesh file.

* Material name

Prepare for adding the material name to an element if there are interface boundaries.

* Change node pointer storage in elements

Instead of using an objectArray to store the pointers for the 4/8 element nodes, use an allocatable array of SMNodePtrs. This seems to avoid not the memory crash when using interface boundaries.

* Setting material IDs

Move the setting of material IDs to the very end. Delete some unused code.

* Change material setting logic

Material properties are only set if the ISM_MM output format is requested.

* Update MeshGeneratorMethods.f90

Remove redundant material setting code now covered in SetMaterialProperties

* Fix generic exception catches

In several places execution was altered if non-fatal errors were present. Changes general catch to catch only on fatal exceptions.

* Add tests for multi material

BoneAndMarrow.control added to test suite.

* Update benchmark files

The changes made don't generate unused nodes in the mesh. Other than that, the meshes are the same.

* Update docs and examples

Add documentation on interface boundaries. Add BoneAndMarrow example. **The new added figure needs to be moved from the Figures directory to where all the other figures are stored in github and references changed in two places: Introduction and theModel**

* Update Documentation/docs/introduction.md

Fix typo in new addition to the documentation

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>

* Background grid naming fix

Fixes two main issues with output of the mesh in the -MM format.
1. Fix the name of the background material keyword to "background material name" and check if it is set when the material key is present. If not, a warning is posted and the background material name is defaulted to "base"
2. Fix the output of the base material name by removing the "materialName" property from the element definition and printing the material name using the materialID of the element and the materialNameForIDs array held by the mesh.
3. Do #2 for hex meshes, too.

* Cleanup

Fix typos in docs and control files. Remove commented out code that used old data structures.

* update docs with new link to the interior boundaries figure

* remove figure file Documentation/Figures/BoneAndMarrow.png

* update CI actions

* Apply suggestions from code review

* update location of BoneAndMarrow generated files

* fix typo in docs

---------

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>
Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>