Add test with `ifort` compiler (#76)
* Add test with `ifort` compiler

* MoveClasstotype

About half of the code has class changed to type where possible.

* More class->types.

* Casting change

Make cast to element and edge specific.

* Uset type

use type instead of class

* ClassToType

Change node, edge and element class to type since there are no subclasses of those.

* Update FRSegmentedCurveClass.f90

Change newPointArray to type

* Update Sizer.f90

class to type

* Update MeshOperationsModule.f90

change node to TYPE.

* Node pointers

use type for more node pointers missed before.

* Change missed types

Still a few more node and elements needed change to TYPE.

* Make QuadtreeGrid a type

* ConvertSMElementToType

Except for one place where gfortran throws an internal error, convert all class(SMElement) to TYPE(SMElement)

* ConvertNodesAndEdges

convert nodes and edges to TYPE.

* Update SpringMeshSmoother.f90

Make smootherDict class consistent.

* Update SweeperClass.f90

Curves, in general will be subclassed, so use CLASS instead of TYPE.

* SpecificReleases

Use specific (e.g. releaseSMQuad) rather than generic (release(obj)) to work with the intel compiler. This has been a longstanding issue and look for this in the future. It looks best to use the specific release function.

* Update CurveTests.f90

Change the relative tolerance for the curve tests from 1d-8 to 2e-8. The intel compiler passed with the former with -O0, but failed with -O. This change makes both compilers pass with both optimizations.

* add fix for fetching valgrind

* attempt to fix broken ifort runner

* revert change that didn't fix issue

* add that environment is set from cahced version of ifort as well

* add that environment is set from cached version of ifort as well

* add new name line to fix broken script

* Update docs

Add to the readme and documentation that both ifort and gfortran are supported, but that ifort was found to be slower. Copy the more extensive readme contents about compiling, etc. to the compiling section of the docs. Fix a couple of typos and minor edits.

* small style changes in the new docs

---------

Co-authored-by: David Kopriva <kopriva@math.fsu.edu>
Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>