Fix boundary edge/curve error
Fixes a longstanding error in assigning the boundary curve to an element boundary edge. Instead of using ROW_END as a discriminant to choose which curve the two nodes fall on (a corner node can be on two curves simultaneously), it checks to see if the node is at the endpoints of its curve and if so, uses the curve attached to the other node.