Symmetrize (#99)
* Straight curve test

Add routine to see if a curve is a straight line by checking the second derivative at a point. Add tests for that, too. See comments for limitations.

* Start mesh mirroring code.

Start adding code to mirror across a symmetry line.

* AddLineReflections

Add code to compute the reflection of a point about a line, plus tests. Put in stubs for mirroring a mesh.

* Create Symmetric Mesh

Add changes to flip a mesh about a symmetry line and add the new elements to the mesh. Only does one symmetry boundary in the model ATM. Haven't checked the mesh file yet, and edges need to be added still to handle ISM-V2.

* Add Edge reflection and reflect xPatch

Add reflection of the edges in the mesh by removing the old ones and re-computing them. Also reflect the xPatch array, which is used in the SEM plotting procedure. Finally update the makefile to include the new reflection code.

* ifort fixes

Fix ifort error regarding class/type of a variable. Fix typos.

* Symmetry Testing

Add routines to test if all curves marked symmetry are co-linear straight lines. If not, posts a warning and the mesh is not reflected.

* Fix boundary edge/curve error

Fixes a longstanding error in assigning the boundary curve to an element boundary edge. Instead of using ROW_END as a discriminant to choose which curve the two nodes fall on (a corner node can be on two curves simultaneously), it checks to see if the node is at the endpoints of its curve and if so, uses the curve attached to the other node.

* Fix boundary info curves

Make boundary curves int the boundaryInfo consistent with the defined directions.

* add missing link in the Makefile

* Add tests for symmetry reflection

Added PacMan example doing symmetry reflection to test suite. Change test on SYMMETRY_CURVE_NAME to be case insensitive.

* Update documentation

Update the documentation to include the PacMan example.

* Update the-model.md

Update to the previous to include one more file.

* Append to boundary names

Append "_R" to boundary names that are on the reflected portion of the model domain, to allow for different BCs if desired.

* Update Source/Foundation/LineReflectionModule.f90

Add literal type where it was missing.

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>

* remove Reflection figure from repo

* add image url for the new reflection figure

---------

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>