AddGallery (#114)
* AddGallery

Add a Gallery page with plots.

* Update Gallery.md

Fix misspelling

* GalleryReferences

Add references to the gallery in the README and to the menu that shows up on the left. I also rearranged the menu to have the authors and license at the end.

* fix preparedocs

* use gallery images from examples repo

* suggest another formatting and video

* responsive video

* use smaller versions of the images to decrease latency

* Update Gallery.md

Fix multiple the column issue by including html comment markers in for the comments in the div's.

* Add captions to figures in gallery

* Explain how to add new gallery examples

* Update Gallery.md

Change all the div's to be the css versions for two column display plus captions. Change the captions to be of the form `caption (contributor)'. Change the video section name to "Applications".

* Update Gallery.md

Fix a reference/caption

* delete gallery images (moved to examples repo)

* add links to other videos

---------

Co-authored-by: Hendrik Ranocha <mail@ranocha.de>
Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>
Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>