Update HOHQMeshMain.f90 (#121)
* Update HOHQMeshMain.f90

Add help to be printed with the -help option.

* Modify command line behavior

Now accepts -h, -help, --help for help. Prints out help when there is an error in the command line arguments.

* UpdateHelp

Remove -h option. Add comment that partial command line aregumens are handled.

* Update MeshingTests.f90

Add a MiscTests procedure to put in miscellaneous tests from now on. Right now, it calls the PrintHelpMessage procedure to scratch to include covereage.

* Improve test coverage

---------

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>
Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>