Convert to Principal Curvature (#132)
* Convert to Principal Curvature

Replace the gaussian curvature test for topography resolution with the max of the abs of the two principal curvatures. This removes an issue of zero curvature in one of the directions. New mesh benchmarks were needed because the meshes are different with the different curvature tests.

* add license

* Doc Editing and file removal

Edit the doc's for typos discovered along the way, plus some additional notes. Remove files associated with gaussian curvature since they are no longer needed

* typo fix

* point to MIT textbook in the docs for principle curvatures

* Update assumptions-and-algorithms.md

Fix a typo.

* Update appendix.md

Add description of elliptic arc block.

* Create WorkflowTools.md

Add a page that describes two add-ons that can be used to generate control files. Other tools can be added as desired.

* add new docs file to the build

* move the Workflow Tools docs section after the User Guide

---------

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>