Fix description of the gridded bathymetry data format in docs (#134)
* fix description of the gridded bathymetry data format

* Modify Docs about 2D topography sizing

Made a note in the section on refinement regions that automatic sizing can be done with a topography included and created a link. Modified the sizing discussion in the 3D meshing section to say that it can be sued to size 2D meshes.

* Add tl;dr

Added a section to the beginning of the index and the introduction that outlines what can be done with HOHQMesh. [These are the same. Should find out how to "include" a file in markdown to have just one copy]

* Update introduction.md

Fix a typo

* Update README.md

Add the tl;dr to the front page, too.

* Update README.md

Fix links

* Update the-model.md

Fix a formatting error

* Update introduction.md

Fix file names.

* Restructure model definition

Move and add model blocks used in the 3D meshing to the description of the model.

* More doc edits

Make changes sugersted by Andrew, plus a few more, including formatting changes.

---------

Co-authored-by: David Kopriva <kopriva@math.fsu.edu>
Co-authored-by: DavidAKopriva <85404032+DavidAKopriva@users.noreply.github.com>